CM-111: Refactored core/redux/ethereum
