+++ packages/core/redux/ducks.js
@@ -8,8 +8,8 @@ import { combineReducers } from 'redux-immutable'
 import assetsHolder from './assetsHolder/reducer'
 import assetsManager from './assetsManager/reducer'
 import dao from './daos/reducer'
-import ethereum from './ethereum/reducer'
 import ethMultisigWallet from './multisigWallet/reducer'
+import events from './events/reducer'
 import mainWallet from './mainWallet/reducer'
 import market from './market/reducer'
 import notifier from './notifier/reducer'
@@ -19,18 +19,18 @@ import rewards from './rewards/reducer'
 import session from './session/reducer'
 import settingsErc20Tokens from './settings/erc20/tokens/reducer'
 import tokens from './tokens/reducer'
+import txEthereum from './ethereum/reducer'
 import voting from './voting/reducer'
 import wallet from './wallet/reducer'
 import wallets from './wallets/reducer'
 import watcher from './watcher/reducer'
-import events from './events/reducer'
 
 const coreReducers = {
   assetsHolder,
   assetsManager,
   dao,
-  ethereum,
   ethMultisigWallet,
+  events,
   mainWallet,
   market,
   notifier,
@@ -40,11 +40,11 @@ const coreReducers = {
   session,
   settingsErc20Tokens,
   tokens,
+  txEthereum,
   voting,
   wallet,
   wallets,
   watcher,
-  events,
 }
 
 export default coreReducers
