+++ packages/core/redux/session/thunks.js
@@ -18,7 +18,7 @@ import { daoByType } from '../../redux/daos/selectors'
 import web3Factory from '../../web3'
 import { watcher } from '../watcher/actions'
 import { watchStopMarket } from '../market/actions'
-import { initEthereum } from '../ethereum/actions'
+import { initEthereum } from '../ethereum/thunks'
 import { DUCK_PERSIST_ACCOUNT } from '../persistAccount/constants'
 import { DEFAULT_CBE_URL, DEFAULT_USER_URL, DUCK_SESSION } from './constants'
