+++ packages/core/redux/tokens/actions.js
@@ -22,7 +22,7 @@ import Amount from '../../models/Amount'
 import { daoByType } from '../daos/selectors'
 import TxExecModel from '../../models/TxExecModel'
 import { web3Selector } from '../ethereum/selectors'
-import { estimateGas } from '../ethereum/actions'
+import { estimateGas } from '../ethereum/thunks'
 
 import {
   TRANSFER_CANCELLED,
@@ -123,20 +123,22 @@ export const alternateTxHandlingFlow = (dao) => (dispatch) => {
 }
 
 export const initTokens = () => async (dispatch, getState) => {
-  if (getState().get(DUCK_TOKENS).isInited()) {
+  let state = getState()
+  if (state.get(DUCK_TOKENS).isInited()) {
     return
   }
-  const web3 = web3Selector()(getState())
+  const web3 = web3Selector()(state)
   ethereumDAO.connect(web3)
 
   dispatch(alternateTxHandlingFlow(ethereumDAO))
   dispatch(tokensInit())
   dispatch(setTokensFetchingCount(0))
-  const erc20: ERC20ManagerDAO = daoByType('ERC20Manager')(getState())
+  const erc20: ERC20ManagerDAO = daoByType('ERC20Manager')(state)
 
+  state = getState()
   erc20
     .on(EVENT_ERC20_TOKENS_COUNT, async (count) => {
-      const currentCount = getState().get(DUCK_TOKENS).leftToFetch()
+      const currentCount = state.get(DUCK_TOKENS).leftToFetch()
       dispatch(setTokensFetchingCount(currentCount + count + 1 /*eth*/))
 
       // eth
@@ -173,8 +175,9 @@ export const initTokens = () => async (dispatch, getState) => {
 }
 
 export const initBtcLikeTokens = () => async (dispatch, getState) => {
+  const state = getState()
   const btcLikeTokens = [btcDAO, bccDAO, btgDAO, ltcDAO]
-  const currentCount = getState().get(DUCK_TOKENS).leftToFetch()
+  const currentCount = state.get(DUCK_TOKENS).leftToFetch()
   dispatch(setTokensFetchingCount(currentCount + btcLikeTokens.length))
 
   return Promise.all(
@@ -182,7 +185,7 @@ export const initBtcLikeTokens = () => async (dispatch, getState) => {
       .map(async (dao) => {
         try {
           dao.on(EVENT_UPDATE_LAST_BLOCK, (newBlock) => {
-            const blocks = getState().get(DUCK_TOKENS).latestBlocks()
+            const blocks = state.get(DUCK_TOKENS).latestBlocks()
             const currentBlock = blocks[dao.getBlockchain()]
             if (currentBlock && newBlock.block.blockNumber > currentBlock.blockNumber) {
               dispatch(setLatestBlock(newBlock.blockchain, newBlock.block))
